# perf: Improve parser performance for typescript