--- a/packages/babel-parser/src/parse-error/pipeline-operator-errors.ts
+++ b/packages/babel-parser/src/parse-error/pipeline-operator-errors.ts
@@ -12,7 +12,7 @@ type GetSetMemberType<T extends Set<any>> = T extends Set<infer M>
   ? M
   : unknown;
 
-type UnparenthesizedPipeBodyTypes = GetSetMemberType<
+export type UnparenthesizedPipeBodyTypes = GetSetMemberType<
   typeof UnparenthesizedPipeBodyDescriptions
 >;