--- a/packages/babel-parser/src/plugins/flow/scope.ts
+++ b/packages/babel-parser/src/plugins/flow/scope.ts
@@ -1,5 +1,5 @@
 import type { Position } from "../../util/location.ts";
-import ScopeHandler, { Scope } from "../../util/scope.ts";
+import ScopeHandler, { NameType, Scope } from "../../util/scope.ts";
 import { BindingFlag, type ScopeFlag } from "../../util/scopeflags.ts";
 import type * as N from "../../types.ts";
 
@@ -33,11 +33,12 @@ export default class FlowScopeHandler extends ScopeHandler<FlowScope> {
   ): boolean {
     if (super.isRedeclaredInScope(scope, name, bindingType)) return true;
 
-    if (bindingType & BindingFlag.FLAG_FLOW_DECLARE_FN) {
-      return (
-        !scope.declareFunctions.has(name) &&
-        (scope.lexical.has(name) || scope.functions.has(name))
-      );
+    if (
+      bindingType & BindingFlag.FLAG_FLOW_DECLARE_FN &&
+      !scope.declareFunctions.has(name)
+    ) {
+      const type = scope.names.get(name);
+      return (type & NameType.Function) > 0 || (type & NameType.Lexical) > 0;
     }
 
     return false;