… benchmark/babel-parser/real-case-ts/bench.mjs
… benchmark/babel-parser/real-case-ts/ts-parser.txt
… benchmark/babel-parser/real-case/bench.mjs
… benchmark/babel-parser/real-case/jquery-3.6.txt
… packages/babel-parser/src/util/expression-scope.ts
… packages/babel-parser/src/util/scope.ts