@liuxingbaoyu 2023-10-31
[`a25c777` (#16072)](https://github.com/babel/babel/pull/16072/commits/a25c777e7f33b931f52c82567efe4c9c916099ac)
This commit will only have a slight performance impact on errors and will not affect overall performance, but it simplifies our code.